third_party/blink/web_tests/http/tests/xmlhttprequest/xmlhttprequest-overridemimetype-allowedState-expected.txt

Testing overrideMimeType when readyState is 0
PASS xhr.overrideMimeType('text/plain') did not throw exception.
Testing overrideMimeType when readyState is 1
PASS xhr.overrideMimeType('text/plain') did not throw exception.
Testing overrideMimeType when readyState is 1
PASS xhr.overrideMimeType('text/plain') did not throw exception.
Testing overrideMimeType when readyState is 1
PASS xhr.overrideMimeType('text/plain') did not throw exception.
Testing overrideMimeType when readyState is 2
PASS xhr.overrideMimeType('text/plain') did not throw exception.
Testing overrideMimeType when readyState is 3
PASS xhr.overrideMimeType('text/plain') threw exception InvalidStateError: Failed to execute 'overrideMimeType' on 'XMLHttpRequest': MimeType cannot be overridden when the state is LOADING or DONE..
Testing overrideMimeType when readyState is 4
PASS xhr.overrideMimeType('text/plain') threw exception InvalidStateError: Failed to execute 'overrideMimeType' on 'XMLHttpRequest': MimeType cannot be overridden when the state is LOADING or DONE..
PASS successfullyParsed is true

TEST COMPLETE
